feat(server-grpc): add unit tests for gRPC services
