+++ server/src/grpc/mod.rs
@@ -3,6 +3,10 @@
 
 #[macro_use]
 pub mod macros;
+
+#[cfg(test)]
+pub mod tests;
+
 pub mod server;
 
 mod link_service;
@@ -334,481 +338,4 @@ pub fn get_runtime_handle() -> Result<Handle, ArrErr> {
         let rt = Runtime::new()?;
         Ok(rt.handle().clone())
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_util::assert_init_done;
-    use prost_wkt_types::Timestamp;
-    use tonic::Status;
-
-    #[tokio::test]
-    async fn test_from_arrerr_to_status() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        // Create an ArrErr instance with an error message
-        let arr_err = ArrErr::Error("test error message".to_string());
-        // Call the From<ArrErr> for Status implementation to convert the error
-        let status = Status::from(arr_err);
-        // Check that the resulting Status instance has the expected code and message
-        assert_eq!(status.code(), tonic::Code::Internal);
-        assert_eq!(status.message(), "error");
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_bytes() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let bytes = vec![0x68, 0x65, 0x6c, 0x6c, 0x6f];
-
-        // GrpcField into bytes
-        let field = GrpcField::Bytes(bytes.clone());
-        let result: Vec<u8> = field.into();
-        assert_eq!(result, bytes.clone());
-
-        // GrpcFieldOption into bytes
-        let field_option = GrpcFieldOption::Bytes(Some(bytes.clone()));
-        let result: Option<GrpcField> = field_option.into();
-        assert_eq!(result, Some(GrpcField::Bytes(bytes.clone())));
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_string_list() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        // input vec, should return vec
-        let field = GrpcField::StringList(vec!["hello".to_string(), "world".to_string()]);
-        let result = Vec::<String>::from(field);
-        assert_eq!(result, vec!["hello".to_string(), "world".to_string()]);
-
-        // input single string, should return vec
-        let field = GrpcField::String("test".to_string());
-        let result: Vec<String> = field.into();
-        assert_eq!(result, vec!["test".to_string()]);
-
-        // input non string, should return empty list
-        let field = GrpcField::I64(123);
-        let result: Vec<String> = field.into();
-        assert_eq!(result, Vec::<String>::new());
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_string() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let string = String::from("hello");
-
-        // GrpcField into String
-        let field = GrpcField::String(string.clone());
-        let result: String = field.into();
-        assert_eq!(result, string.clone());
-
-        // GrpcFieldOption into String
-        let field_option = GrpcFieldOption::String(Some(string.clone()));
-        let result: Option<GrpcField> = field_option.into();
-        assert_eq!(result, Some(GrpcField::String(string.clone())));
-
-        let field = GrpcFieldOption::String(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::String into String
-        let field = GrpcField::I64(42);
-        let result: String = field.into();
-        assert_eq!(result, "I64(42)");
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_i64_vec() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let i64_vec = vec![1, -2, 3, -4];
-
-        // GrpcField into Vec<i64>
-        let field = GrpcField::I64List(i64_vec.clone());
-        let result: Vec<i64> = field.into();
-        assert_eq!(result, i64_vec.clone());
-
-        // GrpcFieldOption into Vec<i64>
-        let field = GrpcFieldOption::I64List(Some(i64_vec.clone()));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::I64List(i64_vec.clone())));
-
-        let field = GrpcFieldOption::I64List(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // GrpcField::I64 into Vec<i64>
-        let field = GrpcField::I64(42);
-        let result: Vec<i64> = field.into();
-        assert_eq!(result, vec![42]);
-
-        // Non GrpcField::I64List into Vec<i64>
-        let field = GrpcField::Bool(false);
-        let result: Vec<i64> = field.into();
-        assert_eq!(result, Vec::<i64>::new());
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_i64() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let i64 = -42;
-
-        // GrpcField into i64
-        let field = GrpcField::I64(i64);
-        let result: i64 = field.into();
-        assert_eq!(result, i64);
-
-        let field = GrpcField::U32(32);
-        let result: u32 = field.into();
-        assert_eq!(result, 32);
-
-        // GrpcFieldOption into i64
-        let field = GrpcFieldOption::I64(Some(i64));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::I64(i64)));
-
-        let field = GrpcFieldOption::U32(Some(32));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::U32(32)));
-
-        let field = GrpcFieldOption::I64(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::I64 into i64
-        let field = GrpcField::Bool(false);
-        let result: i64 = field.into();
-        assert_eq!(result, 0);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_f64() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let f64 = 42.42;
-
-        // GrpcField into f64
-        let field = GrpcField::F64(f64);
-        let result: f64 = field.into();
-        assert_eq!(result, f64);
-
-        // GrpcFieldOption into f64
-        let field = GrpcFieldOption::F64(Some(f64));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::F64(f64)));
-
-        let field = GrpcFieldOption::F64(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::F64 into f64
-        let field = GrpcField::Bool(false);
-        let result: f64 = field.into();
-        assert_eq!(result, 0.0);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_i32() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let i32 = -42;
-
-        // GrpcField into i32
-        let field = GrpcField::I32(i32);
-        let result: i32 = field.into();
-        assert_eq!(result, i32);
-
-        // GrpcFieldOption into i32
-        let field = GrpcFieldOption::I32(Some(i32));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::I32(i32)));
-
-        let field = GrpcFieldOption::I32(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::I32 into i32
-        let field = GrpcField::Bool(false);
-        let result: i32 = field.into();
-        assert_eq!(result, 0);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_u32() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let u32 = 42;
-
-        // GrpcField into u32
-        let field = GrpcField::U32(u32);
-        let result: u32 = field.into();
-        assert_eq!(result, u32);
-
-        // GrpcFieldOption into u32
-        let field = GrpcFieldOption::U32(Some(u32));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::U32(u32)));
-
-        let field = GrpcFieldOption::U32(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::U32 into u32
-        let field = GrpcField::Bool(false);
-        let result: u32 = field.into();
-        assert_eq!(result, 0);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_u32_vec() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let u32_vec = vec![1, 2, 3];
-
-        // GrpcField into Vec<u32>
-        let field = GrpcField::U32List(u32_vec.clone());
-        let result: Vec<u32> = field.into();
-        assert_eq!(result, u32_vec.clone());
-
-        // GrpcFieldOption into Vec<u32>
-        let field = GrpcFieldOption::U32List(Some(u32_vec.clone()));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::U32List(u32_vec.clone())));
-
-        let field = GrpcFieldOption::U32List(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // GrpcField::U32 into Vec<u32>
-        let field = GrpcField::U32(42);
-        let result: Vec<u32> = field.into();
-        assert_eq!(result, vec![42]);
-
-        // Non GrpcField::U32List into Vec<u32>
-        let field = GrpcField::Bool(false);
-        let result: Vec<u32> = field.into();
-        assert_eq!(result, Vec::<u32>::new());
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_f32() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let f32 = 42.42;
-
-        // GrpcField into f32
-        let field = GrpcField::F32(f32);
-        let result: f32 = field.into();
-        assert_eq!(result, f32);
-
-        // GrpcFieldOption into f32
-        let field = GrpcFieldOption::F32(Some(f32));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::F32(f32)));
-
-        let field = GrpcFieldOption::F32(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::F32 into f32
-        let field = GrpcField::Bool(false);
-        let result: f32 = field.into();
-        assert_eq!(result, 0.0);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_bool() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let bool = true;
-
-        // GrpcField into bool
-        let field = GrpcField::Bool(bool);
-        let result: bool = field.into();
-        assert_eq!(result, bool);
-
-        // GrpcFieldOption into bool
-        let field = GrpcFieldOption::Bool(Some(bool));
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, Some(GrpcField::Bool(bool)));
-
-        let field = GrpcFieldOption::Bool(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        // Non GrpcField::Bool into bool
-        let field = GrpcField::I64(42);
-        let result: bool = field.into();
-        assert_eq!(result, false);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_timestamp() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let timestamp = Timestamp::from(SystemTime::now());
-        let field = GrpcField::Timestamp(timestamp.clone());
-        assert_eq!(timestamp, Timestamp::from(field));
-
-        let timestamp = Timestamp::from(SystemTime::UNIX_EPOCH);
-        let field = GrpcField::Timestamp(timestamp.clone());
-        assert_eq!(timestamp, Timestamp::from(field));
-
-        let field = GrpcField::Bool(false);
-        let result: Timestamp = field.into();
-
-        // this one is tricky as the Timestamp returned from the Bool conversion should be the current timestamp (fallback)
-        // But if we make the comparison with a newly created timestamp, the nanos will be different.
-        // We'll be checking the seconds for now, but this might result in false negatives if the test runs on a second switch.
-        assert_eq!(result.seconds, Timestamp::from(SystemTime::now()).seconds);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_point() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let point = GeoPointZ {
-            x: 120.8,
-            y: 45.12,
-            z: 10.0,
-        };
-
-        // GrpcField into GeoPointZ
-        let field = GrpcField::GeoPointZ(point.clone());
-        let result: GeoPointZ = field.into();
-        assert_eq!(result, point.clone());
-
-        // GrpcFieldOption into GeoPointZ
-        let field_option = GrpcFieldOption::GeoPointZ(Some(point.clone()));
-        let result: Option<GrpcField> = field_option.into();
-        assert_eq!(result, Some(GrpcField::GeoPointZ(point.clone())));
-
-        let field = GrpcFieldOption::GeoPointZ(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_linestring() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let line_string = GeoLineStringZ {
-            points: vec![GeoPointZ {
-                x: 0.12,
-                y: 1.23,
-                z: 4.57,
-            }],
-        };
-
-        // GrpcField into GeoLineStringZ
-        let field = GrpcField::GeoLineStringZ(line_string.clone());
-        let result: GeoLineStringZ = field.into();
-        assert_eq!(result, line_string.clone());
-
-        // GrpcFieldOption into GeoLineStringZ
-        let field_option = GrpcFieldOption::GeoLineStringZ(Some(line_string.clone()));
-        let result: Option<GrpcField> = field_option.into();
-        assert_eq!(result, Some(GrpcField::GeoLineStringZ(line_string.clone())));
-
-        let field = GrpcFieldOption::GeoLineStringZ(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        ut_info!("success");
-    }
-
-    #[tokio::test]
-    async fn test_from_grpc_field_to_polygon() {
-        assert_init_done().await;
-        ut_info!("start");
-
-        let ring_1 = GeoLineStringZ {
-            points: vec![GeoPointZ {
-                x: 0.12,
-                y: 1.23,
-                z: 2.34,
-            }],
-        };
-
-        let ring_2 = GeoLineStringZ {
-            points: vec![
-                GeoPointZ {
-                    x: 0.11,
-                    y: 1.22,
-                    z: 2.35,
-                },
-                GeoPointZ {
-                    x: 0.11,
-                    y: 1.21,
-                    z: 2.36,
-                },
-            ],
-        };
-
-        let polygon = GeoPolygonZ {
-            rings: vec![ring_1, ring_2],
-        };
-
-        // GrpcField into Polygon
-        let field = GrpcField::GeoPolygonZ(polygon.clone());
-        let result: GeoPolygonZ = field.into();
-        assert_eq!(result, polygon.clone());
-
-        // GrpcFieldOption into Polygon
-        let field_option = GrpcFieldOption::GeoPolygonZ(Some(polygon.clone()));
-        let result: Option<GrpcField> = field_option.into();
-        assert_eq!(result, Some(GrpcField::GeoPolygonZ(polygon.clone())));
-
-        let field = GrpcFieldOption::GeoPolygonZ(None);
-        let result: Option<GrpcField> = field.into();
-        assert_eq!(result, None);
-
-        ut_info!("success");
-    }
 }
